added select functionality to location cards
